+++ Src/views/lib/VwGraphics.cpp
@@ -1197,8 +1197,8 @@ STDMETHODIMP VwGraphics::SetupGraphics(LgCharRenderProps * pchrp)
 {
 	BEGIN_COM_METHOD;
 	ChkComArgPtr(pchrp);
-
-	const int cbFontOffset = offsetof(LgCharRenderProps, ttvBold);
+	//offsetof returns a value that is always in the integer range
+	const int cbFontOffset = (int)offsetof(LgCharRenderProps, ttvBold);
 	// if the info related to choosing HFONT is different, make a new HFONT
 	if (!m_hfont || memcmp(((byte *)pchrp) + cbFontOffset, ((byte *)&m_chrp) + cbFontOffset,
 		isizeof(m_chrp) - cbFontOffset))
